Add clipping.

+++ gfx/blit.h
@@ -4,7 +4,7 @@
 #include "gfx/common.h"
 #include "gfx/pixbuf.h"
 
-void PixBufBlit(PixBufT *dstBuf, size_t x, size_t y,
+void PixBufBlit(PixBufT *dstBuf, int x, int y,
                 PixBufT *srcBuf, const RectT *srcRect);
 
 void PixBufBlitScaled(PixBufT *dstBuf, size_t x, size_t y, int w, int h,
